calculatie/gRPC_server/app/requirements.txt
protobuf==5.29.2
grpcio==1.68.1
grpcio-tools==1.68.1
pymysql
